# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API with Swagger",
        "description": "RESTFULL API sistem penjualan documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/transaction": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Service Transaction",
                "description": "Endpoint get transaction",
                "operationId": "74e9cda359a93b89d1502508ad10aee9",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Service Transaction",
                "description": "Endpoint post transaction",
                "operationId": "9e3072adb52a6e629183961c7919ba78",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "product_id field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "qty",
                        "in": "query",
                        "description": "qty field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "total",
                        "in": "query",
                        "description": "total field",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    },
                    "400": {
                        "description": "Field product must be filled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/transaction/{transaction_number}": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Service transaction",
                "description": "Endpoint get transaction specified id",
                "operationId": "35f132a741cbed66ad0aac274bb33522",
                "parameters": [
                    {
                        "name": "transaction_number",
                        "in": "path",
                        "description": "transaction transaction_number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Service transaction",
                "description": "Endpoint add new transaction",
                "operationId": "2e5abd0740706fd0278b8c1b59e7d30d",
                "parameters": [
                    {
                        "name": "transaction_number",
                        "in": "path",
                        "description": "category data transaction_number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "product_id field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "qty",
                        "in": "query",
                        "description": "qty field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "total",
                        "in": "query",
                        "description": "total field",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    },
                    "400": {
                        "description": "Field product must be filled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Delete existing transaction",
                "description": "Deletes a record and returns no content",
                "operationId": "delete transaction",
                "parameters": [
                    {
                        "name": "transaction_number",
                        "in": "path",
                        "description": "product transaction_number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}